🔥 chore: move from the recoil folder to the types folder
